chore(trading-competition-end-phase): Move phases into config and tidy 'types'
