+++ src/views/TradingCompetition/components/BattleCta/index.tsx
@@ -44,6 +44,7 @@ const StyledButton = styled(Button)`
 
 const BattleCta: React.FC<CompetitionProps> = ({
   userTradingInformation,
+  currentPhase,
   account,
   isCompetitionLive,
   profile,
