+++ src/views/TradingCompetition/components/Countdown/index.tsx
@@ -3,8 +3,9 @@ import styled from 'styled-components'
 import { Flex, Skeleton, PocketWatchIcon, Text } from '@pancakeswap-libs/uikit'
 import useI18n from 'hooks/useI18n'
 import getTimePeriods from 'utils/getTimePeriods'
+import { CompetitionSteps, FINISHED, LIVE } from 'config/constants/trading-competition/easterPhases'
 import { Heading2Text } from '../CompetitionHeadingText'
-import { CompetitionSteps, FINISHED, LIVE, CompetitionPhaseProps } from '../../config'
+import { CompetitionPhaseProps } from '../../types'
 import ProgressStepper from './ProgressStepper'
 import Timer from '../../../../components/Timer'
 import { GOLDGRADIENT } from '../Section/sectionStyles'
