+++ src/views/TradingCompetition/types.tsx
@@ -1,20 +1,42 @@
 import { ReactText } from 'react'
 import { Profile } from 'state/types'
 
-export interface UserRewardsProps {
-  userCanClaim?: boolean
-  userRewards?: {
-    cakeToClaim?: string
-    pointsToClaim?: string
-  }
+export interface CompetitionProps extends UserRewardsProps {
+  userTradingInformation?: UserTradingInformationProps
+  currentPhase?: CompetitionPhaseProps
+  account?: string
+  profile?: Profile
+  isCompetitionLive?: boolean
+  hasCompetitionFinished?: boolean
+  isLoading?: boolean
+  onDismiss?: () => void
+  onRegisterSuccess?: () => void
+  onClaimSuccess?: () => void
 }
-export interface UserTradingInformationProps {
-  hasRegistered?: boolean
-  hasUserClaimed?: boolean
-  userRewardGroup?: string
-  userCakeRewards?: string
-  userPointReward?: string
-  canClaimNFT?: boolean
+
+export interface CompetitionStepProps {
+  index?: number
+  text?: string
+  translationId: number
+}
+
+export interface CompetitionPhaseProps {
+  state?: string
+  ends?: number | null
+  step?: CompetitionStepProps
+}
+
+export interface CountdownProps {
+  steps?: Array<{ text: string; translationId: number }>
+  activeStepIndex?: number
+  stepText?: string
+  index?: number
+}
+
+interface LeaderboardData {
+  total?: number
+  volume?: number
+  data?: LeaderboardDataItem[]
 }
 
 export interface LeaderboardDataItem {
@@ -24,10 +46,21 @@ export interface LeaderboardDataItem {
   teamId?: number
 }
 
-interface LeaderboardData {
-  total?: number
-  volume?: number
-  data?: LeaderboardDataItem[]
+export interface RibbonProps {
+  ribbonDirection?: 'up' | 'down'
+  ribbonText?: string
+  isCardHeader?: boolean
+  imageComponent?: React.ReactNode
+  children?: React.ReactNode
+}
+
+export interface SectionProps {
+  backgroundStyle?: string
+  svgFill?: string
+  index?: number
+  intersectionPosition?: 'top' | 'bottom'
+  intersectComponent?: React.ReactNode
+  noIntersection?: boolean
 }
 
 export interface TeamLeaderboardProps {
@@ -43,16 +76,21 @@ export interface TeamRanksProps {
   isGlobalLeaderboardDataComplete?: boolean
 }
 
-export interface CompetitionProps extends UserRewardsProps {
-  userTradingInformation?: UserTradingInformationProps
-  account?: string
-  profile?: Profile
-  isCompetitionLive?: boolean
-  hasCompetitionFinished?: boolean
-  isLoading?: boolean
-  onDismiss?: () => void
-  onRegisterSuccess?: () => void
-  onClaimSuccess?: () => void
+export interface UserRewardsProps {
+  userCanClaim?: boolean
+  userRewards?: {
+    cakeToClaim?: string
+    pointsToClaim?: string
+  }
+}
+
+export interface UserTradingInformationProps {
+  hasRegistered?: boolean
+  hasUserClaimed?: boolean
+  userRewardGroup?: string
+  userCakeRewards?: string
+  userPointReward?: string
+  canClaimNFT?: boolean
 }
 
 export interface YourScoreProps extends CompetitionProps {
@@ -64,26 +102,4 @@ export interface YourScoreProps extends CompetitionProps {
     // eslint-disable-next-line camelcase
     next_rank?: number
   }
-}
-export interface RibbonProps {
-  ribbonDirection?: 'up' | 'down'
-  ribbonText?: string
-  isCardHeader?: boolean
-  imageComponent?: React.ReactNode
-  children?: React.ReactNode
-}
-export interface SectionProps {
-  backgroundStyle?: string
-  svgFill?: string
-  index?: number
-  intersectionPosition?: 'top' | 'bottom'
-  intersectComponent?: React.ReactNode
-  noIntersection?: boolean
-}
-
-export interface CountdownProps {
-  steps?: Array<{ text: string; translationId: number }>
-  activeStepIndex?: number
-  stepText?: string
-  index?: number
 }
